Don't confirm import of undated files.
The slowdown of confirming an image has no known date isn't worthwhile
if a large number of files fall under that category. The scenario of
accidental import that this was meant to catch can otherwise be
handled by the deport command.

Also remove the related flag --with-certainty, which was similarly
less useful than hoped for due to the extra typing it entails.