Overhaul database schema to reduce redundancy.
The metadata table is now the archive table. The fts table is now the
archive_fts table.

The path field has been dropped from the archive table, since its
value is already captured in the JSON of the tags field. Make that
value relative to the archive root rather than absolute for greater
portability.

Also adds a reindex command to help with recreating the database from
scratch.