Redo keyword untagging.
It's no longer convenient to pre-check aliases against the config file
using parameter validation. This in turn complicates the use of multi
methods for untag operations specified in different ways.

Instead, use a second parameter to describe the kind of untagging
being done and offload most of the work to a module.