Add support for order search results by series id.
Expand search command to allow for ordering to be specified in a
similar manner to faceted keywords. Also add a --debug flag to display
the SQL used to perform the search. When displaying results, include
the series name and number.

This requires a database reindex to switch to flatter JSON when
exporting tags from exiftool.