Set permissions to read-write when copying into a workspace.
Don't try to keep v-00 read-only and everything else read-write
because it interferes with manual editing workflows in non-obvious
ways (such as copying v-00 to v-01 and then having to remember to
chmod v-01). Only set read-only when a file comes into the archive or
is promoted out of a workspace, and rely on filesystem backups to
handle accidents.