Don't error during reprompt when file doesn't move.
If an archived file was retagged via the reprompt command, an error
would occur from an unnecessary move operation. Reorder the precision
and ordering of the check for whether the source and destination of a
file are the same during import. Also make the caller responsible for
reindexing instead of handling during import for better consistency.